fix(education): Resolve all build errors and add missing implementations
- Add education source files to Makefile (education_db, setup_wizard,
  anna_integration, features/*, tools/*)
- Add education object directories to build system
- Fix EducationAccessibility struct: add adhd_severity, tts_voice,
  high_contrast, reduce_motion fields
- Fix extern declarations in study_session.c to match header signatures
- Fix anna_integration.c: use adhd_severity instead of bool for severity
- Fix quiz.c: add unistd.h, fix cp_severity -> cerebral_palsy_severity
- Fix flashcards.c: add unistd.h for getpid/unlink
- Fix calculator.c: add forward declaration for calc_solve_linear
- Export g_edu_db for feature modules
- Implement goal management functions: education_goal_add,
  education_goal_list, education_goal_achieve, education_goal_delete,
  education_goal_list_free
- Implement education_maestro_broadcast_profile stub
- Add llm_generate weak stub for feature modules

Build now completes successfully with 0 errors.

+++ src/education/features/study_session.c
@@ -32,8 +32,8 @@
 
 extern sqlite3* g_edu_db;
 extern char* llm_generate(const char* prompt, const char* system_prompt);
-extern int education_session_start(int64_t student_id, const char* session_type,
-                                   const char* subject, const char* topic);
+extern int64_t education_session_start(int64_t student_id, const char* session_type,
+                                       const char* subject, const char* topic);
 extern int education_session_end(int64_t session_id, int xp_earned);
 extern int education_xp_add(int64_t student_id, int xp_amount, const char* reason);
